restructure module
