+++ src/core/main/java/org/comroid/mcsd/core/module/AbstractModule.java
@@ -4,12 +4,19 @@ import lombok.Data;
 import org.comroid.api.Component;
 import org.comroid.api.Named;
 import org.comroid.mcsd.core.entity.Server;
+import org.jetbrains.annotations.Nullable;
 
 @Data
-public abstract class AbstractModule extends Component.Base implements Named, Runnable {
-    private final Server server;
+public abstract class AbstractModule extends Component.Base implements Named {
+    protected final Server server;
 
     @Override
-    public void run() {
+    public final @Nullable Component getParent() {
+        return server;
+    }
+
+    @Override
+    public final Component.Base setParent(@Nullable Component parent) {
+        throw new UnsupportedOperationException();
     }
 }
