--- src/core/main/java/org/comroid/mcsd/core/module/server/DiscordModule.java
@@ -1,10 +1,0 @@
-package org.comroid.mcsd.core.module.server;
-
-import org.comroid.mcsd.core.entity.Server;
-import org.comroid.mcsd.core.module.AbstractModule;
-
-public abstract class DiscordModule extends AbstractModule {
-    public DiscordModule(Server server) {
-        super(server);
-    }
-}
